refactor: use global property type context
