+++ apps/hash-frontend/src/pages/[shortname]/types/entity-type/[...slug-maybe-version].page.tsx
@@ -41,7 +41,6 @@ import { EntityTypeContext } from "./[...slug-maybe-version].page/shared/entity-
 import { EntityTypeEntitiesContext } from "./[...slug-maybe-version].page/shared/entity-type-entities-context";
 import { EntityTypeHeader } from "./[...slug-maybe-version].page/shared/entity-type-header";
 import { getEntityTypeBaseUrl } from "./[...slug-maybe-version].page/shared/get-entity-type-base-url";
-import { LatestPropertyTypesContextProvider } from "./[...slug-maybe-version].page/shared/latest-property-types-context";
 import { useCurrentTab } from "./[...slug-maybe-version].page/shared/tabs";
 import { TypePreviewSlide } from "./[...slug-maybe-version].page/type-preview-slide";
 import { useEntityTypeEntitiesContextValue } from "./[...slug-maybe-version].page/use-entity-type-entities-context-value";
@@ -224,142 +223,138 @@ const Page: NextPageWithLayout = () => {
     <>
       <NextSeo title={`${entityType.title} | Entity Type`} />
       <EntityTypeFormProvider {...formMethods}>
-        <LatestPropertyTypesContextProvider>
-          <EntityTypeContext.Provider value={entityType}>
-            <EntityTypeEntitiesContext.Provider value={entityTypeEntitiesValue}>
-              <Box display="contents" component="form" onSubmit={handleSubmit}>
-                <TopContextBar
-                  defaultCrumbIcon={null}
-                  item={remoteEntityType ?? undefined}
-                  crumbs={[
-                    {
-                      title: "Types",
-                      id: "types",
-                    },
-                    {
-                      title: `${entityTypeIsLink ? "Link" : "Entity"} Types`,
-                      id: "entity-types",
-                    },
-                    {
-                      title: entityType.title,
-                      href: "#",
-                      id: entityType.$id,
-                      icon: <FontAwesomeIcon icon={faAsterisk} />,
-                    },
-                  ]}
-                  scrollToTop={() => {}}
-                  sx={{ bgcolor: "white" }}
-                />
+        <EntityTypeContext.Provider value={entityType}>
+          <EntityTypeEntitiesContext.Provider value={entityTypeEntitiesValue}>
+            <Box display="contents" component="form" onSubmit={handleSubmit}>
+              <TopContextBar
+                defaultCrumbIcon={null}
+                item={remoteEntityType ?? undefined}
+                crumbs={[
+                  {
+                    title: "Types",
+                    id: "types",
+                  },
+                  {
+                    title: `${entityTypeIsLink ? "Link" : "Entity"} Types`,
+                    id: "entity-types",
+                  },
+                  {
+                    title: entityType.title,
+                    href: "#",
+                    id: entityType.$id,
+                    icon: <FontAwesomeIcon icon={faAsterisk} />,
+                  },
+                ]}
+                scrollToTop={() => {}}
+                sx={{ bgcolor: "white" }}
+              />
 
-                {!isReadonly && (
-                  <EditBarTypeEditor
-                    currentVersion={currentVersion}
-                    discardButtonProps={
-                      // @todo confirmation of discard when draft
-                      isDraft
-                        ? {
-                            href: `/new/types/entity-type`,
-                          }
-                        : {
-                            onClick() {
-                              reset();
-                            },
-                          }
-                    }
-                  />
-                )}
-
-                <Box
-                  sx={{
-                    borderBottom: 1,
-                    borderColor: "gray.20",
-                    pt: 3.75,
-                    backgroundColor: "white",
-                  }}
-                >
-                  <Container>
-                    <Box
-                      display="flex"
-                      justifyContent="space-between"
-                      alignItems="flex-start"
-                    >
-                      <EntityTypeHeader
-                        ontologyChip={
-                          <OntologyChip
-                            icon={<OntologyIcon />}
-                            domain="hash.ai"
-                            path={
-                              <>
-                                <Typography
-                                  component="span"
-                                  fontWeight="bold"
-                                  color={(theme) => theme.palette.blue[70]}
-                                >
-                                  {router.query.shortname}
-                                </Typography>
-                                <Typography
-                                  component="span"
-                                  color={(theme) => theme.palette.blue[70]}
-                                >
-                                  /types/entity-type/
-                                </Typography>
-                                <Typography
-                                  component="span"
-                                  fontWeight="bold"
-                                  color={(theme) => theme.palette.blue[70]}
-                                >
-                                  {slug}
-                                </Typography>
-                                <Typography
-                                  component="span"
-                                  color={(theme) => theme.palette.blue[70]}
-                                >
-                                  /v/{currentVersion}
-                                </Typography>
-                              </>
-                            }
-                          />
+              {!isReadonly && (
+                <EditBarTypeEditor
+                  currentVersion={currentVersion}
+                  discardButtonProps={
+                    // @todo confirmation of discard when draft
+                    isDraft
+                      ? {
+                          href: `/new/types/entity-type`,
                         }
-                        entityType={entityType}
-                        isReadonly={isReadonly}
-                        latestVersion={latestVersion}
-                      />
-                      {!isReadonly && !isDraft && !entityTypeIsLink ? (
-                        <ConvertTypeButton
-                          onClick={convertToLinkType}
-                          loading={convertTypeLoading}
-                          disabled={isDirty}
-                        />
-                      ) : null}
-                    </Box>
+                      : {
+                          onClick() {
+                            reset();
+                          },
+                        }
+                  }
+                />
+              )}
 
-                    <EntityTypeTabs isDraft={isDraft} />
-                  </Container>
-                </Box>
-
-                <Box py={5}>
-                  <Container>
-                    {currentTab === "definition" ? (
-                      entityTypeAndPropertyTypes ? (
-                        <DefinitionTab
-                          entityTypeAndPropertyTypes={
-                            entityTypeAndPropertyTypes
+              <Box
+                sx={{
+                  borderBottom: 1,
+                  borderColor: "gray.20",
+                  pt: 3.75,
+                  backgroundColor: "white",
+                }}
+              >
+                <Container>
+                  <Box
+                    display="flex"
+                    justifyContent="space-between"
+                    alignItems="flex-start"
+                  >
+                    <EntityTypeHeader
+                      ontologyChip={
+                        <OntologyChip
+                          icon={<OntologyIcon />}
+                          domain="hash.ai"
+                          path={
+                            <>
+                              <Typography
+                                component="span"
+                                fontWeight="bold"
+                                color={(theme) => theme.palette.blue[70]}
+                              >
+                                {router.query.shortname}
+                              </Typography>
+                              <Typography
+                                component="span"
+                                color={(theme) => theme.palette.blue[70]}
+                              >
+                                /types/entity-type/
+                              </Typography>
+                              <Typography
+                                component="span"
+                                fontWeight="bold"
+                                color={(theme) => theme.palette.blue[70]}
+                              >
+                                {slug}
+                              </Typography>
+                              <Typography
+                                component="span"
+                                color={(theme) => theme.palette.blue[70]}
+                              >
+                                /v/{currentVersion}
+                              </Typography>
+                            </>
                           }
-                          onNavigateToType={onNavigateToType}
-                          ownedById={routeNamespace.accountId as OwnedById}
-                          readonly={isReadonly}
                         />
-                      ) : (
-                        "Loading..."
-                      )
+                      }
+                      entityType={entityType}
+                      isReadonly={isReadonly}
+                      latestVersion={latestVersion}
+                    />
+                    {!isReadonly && !isDraft && !entityTypeIsLink ? (
+                      <ConvertTypeButton
+                        onClick={convertToLinkType}
+                        loading={convertTypeLoading}
+                        disabled={isDirty}
+                      />
                     ) : null}
-                    {currentTab === "entities" ? <EntitiesTab /> : null}
-                  </Container>
-                </Box>
+                  </Box>
+
+                  <EntityTypeTabs isDraft={isDraft} />
+                </Container>
               </Box>
-            </EntityTypeEntitiesContext.Provider>
-          </EntityTypeContext.Provider>
-        </LatestPropertyTypesContextProvider>
+
+              <Box py={5}>
+                <Container>
+                  {currentTab === "definition" ? (
+                    entityTypeAndPropertyTypes ? (
+                      <DefinitionTab
+                        entityTypeAndPropertyTypes={entityTypeAndPropertyTypes}
+                        onNavigateToType={onNavigateToType}
+                        ownedById={routeNamespace.accountId as OwnedById}
+                        readonly={isReadonly}
+                      />
+                    ) : (
+                      "Loading..."
+                    )
+                  ) : null}
+                  {currentTab === "entities" ? <EntitiesTab /> : null}
+                </Container>
+              </Box>
+            </Box>
+          </EntityTypeEntitiesContext.Provider>
+        </EntityTypeContext.Provider>
       </EntityTypeFormProvider>
 
       {previewEntityTypeUrl ? (
