+++ apps/hash-frontend/src/pages/[shortname]/types/entity-type/[...slug-maybe-version].page/definition-tab.tsx
@@ -5,9 +5,9 @@ import { OwnedById, PropertyTypeWithMetadata } from "@local/hash-subgraph";
 import { useMemo } from "react";
 
 import { useEntityTypesContextRequired } from "../../../../../shared/entity-types-context/hooks/use-entity-types-context-required";
+import { useLatestPropertyTypes } from "../../../../../shared/latest-property-types-context";
 import { useEditorOntologyFunctions } from "./definition-tab/use-editor-ontology-functions";
 import { getTypesWithoutMetadata } from "./shared/get-types-without-metadata";
-import { useLatestPropertyTypes } from "./shared/latest-property-types-context";
 
 type DefinitionTabProps = {
   ownedById?: OwnedById;
