+++ apps/hash-frontend/src/pages/[shortname]/types/entity-type/[...slug-maybe-version].page/definition-tab/use-editor-ontology-functions.ts
@@ -19,10 +19,10 @@ import { useBlockProtocolGetPropertyType } from "../../../../../../components/ho
 import { useBlockProtocolUpdateEntityType } from "../../../../../../components/hooks/block-protocol-functions/ontology/use-block-protocol-update-entity-type";
 import { useBlockProtocolUpdatePropertyType } from "../../../../../../components/hooks/block-protocol-functions/ontology/use-block-protocol-update-property-type";
 import { useFetchEntityTypes } from "../../../../../../shared/entity-types-context/hooks";
+import { useFetchLatestPropertyTypes } from "../../../../../../shared/latest-property-types-context";
 import { canUserEditResource } from "../../../../../../shared/readonly-mode";
 import { useAuthInfo } from "../../../../../shared/auth-info-context";
 import { useGenerateTypeUrlsForUser } from "../../../../../shared/use-generate-type-urls-for-user";
-import { useFetchLatestPropertyTypes } from "../shared/latest-property-types-context";
 
 export const useEditorOntologyFunctions = (
   ownedById: OwnedById | null,
