+++ apps/hash-frontend/src/shared/latest-property-types-context.tsx
@@ -2,7 +2,7 @@ import { VersionedUrl } from "@blockprotocol/type-system/slim";
 import { PropertyTypeWithMetadata } from "@local/hash-subgraph";
 import { createContext, PropsWithChildren, useContext } from "react";
 
-import { useLatestPropertyTypesContextValue } from "./use-latest-property-types-context-value";
+import { useLatestPropertyTypesContextValue } from "./latest-property-types-context/use-latest-property-types-context-value";
 
 export type LatestPropertyTypesContextValues = {
   propertyTypes: Record<VersionedUrl, PropertyTypeWithMetadata> | null;
