+++ apps/hash-frontend/src/shared/latest-property-types-context/use-latest-property-types-context-value.ts
@@ -2,8 +2,8 @@ import { PropertyTypeRootType } from "@local/hash-subgraph";
 import { getRoots } from "@local/hash-subgraph/stdlib";
 import { useCallback, useEffect, useMemo, useState } from "react";
 
-import { useBlockProtocolQueryPropertyTypes } from "../../../../../../components/hooks/block-protocol-functions/ontology/use-block-protocol-query-property-types";
-import { LatestPropertyTypesContextValues } from "./latest-property-types-context";
+import { useBlockProtocolQueryPropertyTypes } from "../../components/hooks/block-protocol-functions/ontology/use-block-protocol-query-property-types";
+import { LatestPropertyTypesContextValues } from "../latest-property-types-context";
 
 export const useLatestPropertyTypesContextValue = () => {
   const [propertyTypes, setPropertyTypes] = useState<
